+++ HISTORY.rst
@@ -28,6 +28,7 @@ Unreleased
 * ALCN-1455: Fix new sale order exported with status empty in xml (ESB)
 * ALCN-1406: Internal Transfer: Do not compute product putaway on fixed location otherwise destination is always product's locator
 * ALCN-1448: Rename "End of Life" by "Expiration date" and display lots in a product
+* ALCN-1444: Compute promotion when the procurement is running
 
 **Build**